Restructured project with lib and easier main.rs
