+++ src/lib.rs
@@ -1,0 +1,5 @@
+pub mod file;
+pub mod settings;
+
+#[macro_use]
+extern crate serde_derive;
